+++ lib/cli/text.rb
@@ -5,7 +5,7 @@ module Coltrane
     # A text representation
     class Text < Representation
       def render
-        case @flavor
+        case Cli.config.flavor
         when :marks, :notes, :degrees then @notes.pretty_names.join(' ')
         when :intervals then @notes.map { |n| (@notes.first - n).name }.join(' ')
         else raise WrongFlavorError