+++ lib/coltrane_instruments/guitar.rb
@@ -1,12 +1,10 @@
 # frozen_string_literal: true
 
-require 'coltrane_instruments/guitar/base'
-require 'coltrane_instruments/guitar/chord'
-require 'coltrane_instruments/guitar/note'
-require 'coltrane_instruments/guitar/string'
-
 module ColtraneInstruments
-  # This module  represents a guitar
   module Guitar
+    autoload :Base,   'coltrane_instruments/guitar/base'
+    autoload :Chord,  'coltrane_instruments/guitar/chord'
+    autoload :Note,   'coltrane_instruments/guitar/note'
+    autoload :String, 'coltrane_instruments/guitar/string'
   end
 end