+++ lib/coltrane_synth.rb
@@ -1,0 +1,7 @@
+require 'coltrane'
+require 'coreaudio'
+
+module ColtraneSynth
+  autoload :Base,  'coltrane_synth/base'
+  autoload :Synth, 'coltrane_synth/synth'
+end